Restore default Firefox focus style for `.Button` outside
Firefox's default focus style appears inside buttons and submits, which looks
pretty nasty with our style of button. It's removed in `normalize.css`.

This restores it, but puts it outside the element.

Colour is hardcoded, which is a bit nasty, but Firefox doesn't have an
equivalent to WebKit's `-webkit-focus-ring-color` property.